implemented a2c and put rewards_to_go() into utils.py since that code is shared with other algorithms.
