+++ squiRL/common/utils.py
@@ -1,3 +1,5 @@
+import copy
+import torch
 from torch.utils.data._utils import collate
 from squiRL.common.data_stream import Experience
 
@@ -17,3 +19,23 @@ def collate_episodes(batch):
     }
 
     return batch_dict.values()
+
+
+def reward_to_go(self, rewards: torch.Tensor,
+                 states: torch.Tensor) -> torch.tensor:
+    """Calculates reward to go over an entire episode
+
+    Args:
+        rewards (torch.Tensor): Episode rewards
+
+    Returns:
+        torch.tensor: Reward to go for each episode step
+    """
+    rewards = rewards.detach().cpu().numpy()
+    res = []
+    sum_r = 0.0
+    for r in reversed(rewards):
+        sum_r *= self.gamma
+        sum_r += r
+        res.append(copy(sum_r))
+    return list(reversed(res))
